libs/cgse-common/pyproject.toml
[project]
name = "cgse-common"
version = "0.19.2"
description = "Software framework to support hardware testing"
authors = [
    {name = "IvS KU Leuven"}
]
maintainers = [
    {name = "Rik Huygen", email = "rik.huygen@kuleuven.be"},
    {name = "Sara Regibo", email = "sara.regibo@kuleuven.be"}
]
readme = {"file" = "README.md", "content-type" = "text/markdown"}
requires-python = ">=3.10"
license = "MIT"
keywords = [
    "CGSE",
    "Common-EGSE",
    "hardware testing",
    "software framework"
]
dependencies = [
    "click>=8.2.1",
    "deepdiff>=8.1.1",
    "distro>=1.9.0",
    "gitpython>=3.1.44",
    "prometheus-client>=0.21.1",
    "psutil>=6.1.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.16.0",
    "navdict>=0.5.9",
    # Python 3.9 specific dependencies
    "numpy==1.26.4; python_version == '3.9'",
    "pandas==1.5.1; python_version == '3.9'",
    "pyzmq==23.2.1; python_version == '3.9'",
    # Python 3.10+ specific dependencies
    "numpy>=2.1; python_version >= '3.10'",
    "pandas>=2.2.0; python_version >= '3.10'",
    "pyzmq>=25.1.0; python_version >= '3.10'",
    # Python 3.10+ entrypoints interface changed
    "importlib_metadata>=4.6.0; python_version <= '3.11'",
    "influxdb3-python",
    "duckdb>=1.3.1",
    "requests>=2.32.4",
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.14.0",
]

[project.scripts]
cgse = 'cgse_common.cgse:app'

[project.entry-points."cgse.version"]
cgse-common = 'egse.version:get_version_installed'

[project.entry-points."cgse.settings"]
cgse-common = "cgse_common:settings.yaml"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "-ra --cov --cov-branch --cov-report html"
log_cli = true
log_cli_level = "DEBUG"
filterwarnings = [
    "ignore::DeprecationWarning"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

# Some of these omitted files are only temporary omitted until refactored
[tool.coverage.run]
omit = [
    "tests/*",
    "conftest.py",
    "src/egse/ratelimit.py",
    "src/egse/plugins/metrics/duckdb.py",
    "src/egse/plugins/metrics/timescaledb.py",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests",
    "pytest.ini",
    ".gitignore",
    "*poetry*",
    "*setuptools*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/egse", "src/scripts", "src/cgse_common"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I"]
extend-select = ["E", "W"]

[tool.ruff.lint.isort]
force-single-line = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.0",
    "nox",
    "setuptools>=75.8.2",  # needed by PyCharm
]
